+++ Models/DTO/ProductsDTO.cs
@@ -8,7 +8,7 @@
         public decimal Price { get; set; }
         public int Category{ get; set; }
         public string Color { get; set; }
-        public int ColorNo { get; set; }
+        public long ColorNo { get; set; }
         public decimal Quantity { get; set; }
         public string? ProductImage { get; set; }
         public string? ProductDimensions { get; set; }
